Add files via upload
This project uses a Convolutional Neural Network, to classify between images of people with and without masks. The CNN manages to get an accuracy of 96.88% on the test set. Then the stored weights of this CNN are used to classify as mask or no mask, in real-time, using OpenCV. With the webcam capturing the video, the frames are preprocessed and fed to the model to detect whether Mask is worn or not. It can see Multiple faces as well.